Update all semesters (introduces regressions!)
Use the list of available semesters on Stud.IP (which is parsed anyway) to
sequentially select every single semester and check for any updates in that
semester.

Note that this is currently a pretty sledgehammer-like approach, resulting in
significantly higher network traffic and delay during syncs.
We should probably make it non-default and require a commandline flag to
trigger this mode.

The "remove vanished courses" function can't yet deal with multiple semesters
in the database and has therefore been removed from the UI in this commit.

To make it easier to add a potentially very large number of courses (because
we're now looking at multiple semesters), this commit introduces a "add
all"/"add all from this semester"/"add none"/"add none from this semester"
selection (with keys taken from `git add --patch`).